requirements.txt
numpy==1.23.5
pandas==1.5.2
pyreadstat==1.2.1
pyyaml==6.0